sched/rt: Improve RT throttling debugging
Most of the times, RT throttling is hit due to the current task
hogging the CPU. But we have to do the ramdump analysis to
really confirm that the current task is running for long time.
We can help the cause by printing this information in logbuf.
Print the RT period timer info which will tell if throttling
happened because runtime is not replenished.

Change-Id: Ic032b26b40c771a0e6d9419a6c13ab2f32f83a77
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>